修正 jquery.fetch 加载同 url javascript 文件未完成执行的 undefined BUG；增加表单验证组件

+++ WebContent/assets/js/app-manager.js
@@ -10,7 +10,7 @@
 	yi.config("../");
 
 	yi.ready(function() {
-		window.manager = new Manager();
+		window.manager = new Manager(yi);
 
 		// 详情按钮
 		$('[id^=btn_detail_]').each(function(index, element) {
@@ -35,30 +35,57 @@
 			var name = el.data("name");
 			var version = el.data("version");
 			el.bind("click", function() {
-				el.button('loading');
-				yi.mod.deleteRemoteMod(name, version
-					, function(data) {
-						el.button('reset');
-
-						var table = $('#main-table');
-						table.find('tbody tr').each(function(index, element) {
-							if ($(this).attr('id') == name + '-' + version) {
-								$(this).remove();
-								return false;
-							}
-						});
+				yi.confirm('是否确认要删除模组：<span class="text-info">' + name + ' ' + version + '</span>？', function(result) {
+					if (!result) {
+						return;
 					}
-					, function() {
-						el.button('reset');
-					});
+
+					el.button('loading');
+					yi.mod.deleteRemoteMod(name, version
+						, function(data) {
+							el.button('reset');
+
+							if (data.readOnly !== undefined && data.readOnly) {
+								yi.alert('<p class="text-warning">模组 ' + name + ' ' + version + ' 为只读模组，不能执行“删除”操作。</p>');
+								return;
+							}
+
+							var table = $('#main-table');
+							table.find('tbody tr').each(function(index, element) {
+								if ($(this).attr('id') == name + '-' + version) {
+									$(this).remove();
+									return false;
+								}
+							});
+						}
+						, function() {
+							el.button('reset');
+						});
+				});
 			});
         });
 	});
 
-	function Manager() {
+	function Manager(yi) {
+		this.timer = 0;
+		// 监听失败事件
+		yi.mod.addListener(ModEvent.FAILED, this._onModFailed);
 	}
 
 	Manager.prototype.showAbout = function() {
 		yi.alert('<p>模组管理器 v1.0.0</p><small class="text-muted">Author: Xu Jiangwei (xujiangwei@dhcc.com.cn)</small>');
+	}
+
+	Manager.prototype._onModFailed = function(data) {
+		var self = this;
+		if (self.timer > 0)
+			clearTimeout(self.timer);
+
+		self.timer = setTimeout(function() {
+			clearTimeout(self.timer);
+			self.timer = 0;
+
+			yi.alert('<h5>警告</h5><p class="text-danger">丢失 "'+ data.mod.name +'" 模组，无法使用“详情”功能。</p>');
+		}, 1000);
 	}
 })();
